% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASD185}
\alias{ASD185}
\title{ASD185}
\format{
\subsection{\code{ASD185}}{

List of 1
\describe{
}
}
}
\source{
Fu, J.M., Satterstrom, F.K., Peng, M. et al. Rare coding variation provides insight into the genetic architecture and phenotypic context of autism. Nat Genet 54, 1320–1331 (2022).
}
\usage{
ASD185
}
\description{
185 genes associated with autism spectrum disorder detected by Fu et al. in 2022 with FDR ≤ 0.05
}
\keyword{datasets}
