% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asd_all_output_tidy}
\alias{asd_all_output_tidy}
\title{asd_all_output_tidy}
\format{
\subsection{\code{asd_all_output_tidy}}{

List of 1
\describe{
}
}
}
\usage{
asd_all_output_tidy
}
\description{
The tidy output of map_ukbfgsea on filtered UKB phenotypes for ASD gene set
}
\keyword{datasets}
