% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DD477}
\alias{DD477}
\title{DD477}
\format{
\subsection{\code{DD477}}{

List of 1
\describe{
}
}
}
\source{
Fu, J.M., Satterstrom, F.K., Peng, M. et al. Rare coding variation provides insight into the genetic architecture and phenotypic context of autism. Nat Genet 54, 1320–1331 (2022).
}
\usage{
DD477
}
\description{
477 genes associated with developmental delay detected by Fu et al. in 2022 with FDR ≤ 0.05
}
\keyword{datasets}
