% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd_all_output_tidy}
\alias{dd_all_output_tidy}
\title{dd_all_output_tidy}
\format{
\subsection{\code{dd_all_output_tidy}}{

List of 1
\describe{
}
}
}
\usage{
dd_all_output_tidy
}
\description{
The tidy output of map_ukbfgsea on filtered UKB phenotypes for DD gene set
}
\keyword{datasets}
