% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genebass_test_data}
\alias{genebass_test_data}
\title{Genebass test data}
\format{
\subsection{\code{genebass_test_data}}{

A data frame with 224,630 rows and 44 columns

\describe{
}
}
}
\source{
\url{https://app.genebass.org/downloads}
}
\usage{
genebass_test_data
}
\description{
A subset of data from the genebass gene burden results. The test data includes gene-level test results from three phenotypes, Vascular/heart problems diagnosed by doctor, Year of birth, and HDL cholesterol.
}
\keyword{datasets}
