% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ukbfgsea.R
\name{map_ukbfgsea}
\alias{map_ukbfgsea}
\title{Apply Gene Set Enrichment Analysis to Multiple Phenotypes}
\arguments{
\item{input_geneset}{A list containing one gene set for enrichment analysis. The element of the list should be a character vector of gene Ensembl IDs.}

\item{multiple_phenotypes}{A data frame containing multiple phenotypes from Genebass.}

\item{...}{Additional arguments passed to \code{ukbfgsea}.}
}
\value{
A data frame containing the results of \code{ukbfgsea} for each phenotype annotation. This includes enrichment analysis outcomes using two different ranking metrics, and, if specified, enrichment plots.
}
\description{
This function applies fast preranked gene set enrichment analysis (FGSEA) to multiple phenotypes using a provided gene set. It is designed to handle large-scale datasets efficiently.
}
\examples{
\dontrun{
  library(tidyverse)
  library(ukbFGSEA)
  asd_output <- map_ukbfgsea(input_geneset = ASD185, multiple_phenotypes = genebass_test_data)
}
}
