% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NDD664}
\alias{NDD664}
\title{NDD664}
\format{
\subsection{\code{NDD664}}{

List of 1
\describe{
}
}
}
\source{
Fu, J.M., Satterstrom, F.K., Peng, M. et al. Rare coding variation provides insight into the genetic architecture and phenotypic context of autism. Nat Genet 54, 1320–1331 (2022).
}
\usage{
NDD664
}
\description{
664 genes associated with developmental delay and autism spectrum disorder generated through meta-analysis  by Fu et al. in 2022 with FDR ≤ 0.05
}
\keyword{datasets}
