% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndd_all_output_tidy}
\alias{ndd_all_output_tidy}
\title{ndd_all_output_tidy}
\format{
\subsection{\code{ndd_all_output_tidy}}{

List of 1
\describe{
}
}
}
\usage{
ndd_all_output_tidy
}
\description{
The tidy output of map_ukbfgsea on filtered UKB phenotypes for NDD gene set
}
\keyword{datasets}
