% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_map.R
\name{tidy_map}
\alias{tidy_map}
\title{Tidy the Output of \code{map_ukbfgsea}}
\arguments{
\item{input}{A data frame containing the raw results from \code{map_ukbfgsea()}.}
}
\value{
A tidy data frame with separate columns for each phenotype and ranking method.
}
\description{
This function processes the raw output from \code{map_ukbfgsea()} and organizes it into a tidy format. It splits the output into distinct columns for each phenotype and each ranking method, providing a clear and accessible summary of the FGSEA results.
}
\details{
This function is designed to simplify the interpretation of FGSEA results by restructuring the output into a user-friendly format. It ensures that each phenotype and ranking method is clearly delineated, making it easier to analyze and visualize the results.
}
\examples{
\dontrun{
  library(tidyverse)
  library(ukbFGSEA)
  asd_output <- map_ukbfgsea(input_geneset = ASD185, multiple_phenotypes = genebass_test_data)
  asd_output_tidy <- tidy_map(asd_output)
}
}
