% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbfgsea.R
\name{ukbfgsea}
\alias{ukbfgsea}
\title{Apply Fast Preranked Gene Set Enrichment Analysis (FGSEA) to a Single Phenotype in the UK Biobank Genebass Dataset}
\arguments{
\item{input_geneset}{A list containing one gene set for enrichment analysis. The element of the list should be a character vector of gene Ensembl IDs.}

\item{ukb_genebass}{A data frame containing the Genebass gene burden results of a single phenotype annotation.}

\item{plot}{Logical. If \code{TRUE}, generates GSEA enrichment plots. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{fgsea::fgseaMultilevel}.}
}
\value{
A list with the following components:
\describe{
\item{beta_results}{Enrichment analysis results using \code{rank_beta}.}
\item{beta_stats}{The \code{rank_beta} statistics utilized in the analysis.}
\item{sign_results}{Enrichment analysis results using \code{rank_sign}.}
\item{sign_stats}{The \code{rank_sign} statistics utilized in the analysis.}
\item{beta_plot}{The enrichment plot for \code{rank_beta} if \code{plot = TRUE}.}
\item{sign_plot}{The enrichment plot for \code{rank_sign} if \code{plot = TRUE}.}
}
}
\description{
This function applies \code{fgsea::fgseaMultilevel} - an adaptive multilevel splitting Monte Carlo approach, to a single phenotype in the UK Biobank Genebass dataset for a specified input gene set.
}
\examples{
\dontrun{
  library(tidyverse)
  library(ukbFGSEA)
  one_phenotype <- genebass_test_data \%>\% filter(phenocode=="6150", annotation == "pLoF")
  output <- ukbfgsea(input_geneset = ASD185, ukb_genebass = one_phenotype, plot = T)
}
}
